[DROOLS-6870] Refactored GenerateModelMojo  GenerateModelExecutor
